requirements.txt
pytorch
progressbar2
scikit-learn
OpenCV
matplotlib
pytesseract